FIX: handling exit states must be done after worker threads are stopped
Previously the code was handling the exit states and failure conditions
on the worker watchdog. These checks we done after the threads were
notified to stop (but not necessarily stopped).

This patch moves all the exit state handling code (as well as the test
notifications associated with it) to after the moment the threads are
interrupted or their join timeout has was exhausted.